Decouple Navbar from Header, re-write Navbar scroll functionality
chore: rename Nav to Navbar

the previous implementation of the Navbar scroll functionality was
dependent on the user actually scrolling the page (via scroll
wheel), which may not always be the case. This new imperative
solution sets up an event listener for *any* changes in the scroll
position, which will in turn provide better scroll detection.