fix form field focusing + error message issue
Since formik has a nice ErrorMessage built-in feature, it makes sense to
include that and use the handleFocus/handleBlur event handlers. With
these two components working together, the form fields behave correctly
and populate error messages when the field loses focus. However, the
trade-off is that the error messages do not populate while the user is
typing on the first form fill.

Personally, I think the trade-off is fine. Feel free to revert the
commit or rework some of the functionality (especially where styling is
concerned).